increase timeout for CI environments
